c/partition_balancer: moved types implementation to cc file
(cherry picked from commit 3b20ef27f7ac1712bbdfbf7ee002b6e84f612098)
